MB-56973 [1/6]: Avoid DCP Noops in tests if not necessary
We have a number of tests which enable DCP noops. Most of these tests
don't actually care about testing noops, they just enable them as
noops are manditory for recent features like XATTRs and Collections
(this was a stick to "encourage" DCP consumers which didn't already to
enable them...)

As such, the tests can be potentially brittle if there's any change to
when noops are expected to be sent - for example a slow-running test
or a test paused say under the debugger - if the test takes longer
than the default 20s NOOP interval, then "extra" NOOP requests can be
sent.

To simplify such tests where noops are a unnecessary complication, add
support for enabling Noops with an arbitrarily large TX interval -
they are 'enabled' (and so XATTRs et al are usable), but should never
actually be sent.

Update all tests which configure noops to only actually enable them if
they truely need them.

Change-Id: I116814201f8981368783bef667a73075086884eb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195905
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>