Rename guard-rail to system-constraint and add no-retry
And update the description with information that retrying the
operation will most likely not succeed (this might not be the
case as hitting the error may of course race with ns_server
clearing the restriction).

This information _should_ arrive at the client via the cluster
map (setting and clearing) to allow the client to make an
intelligent decision on when it is possible to perform data
mutations to the node (compared to waste network resources and
CPU resources on the server dealing with operations we _know_
would fail).

Change-Id: I3b37be0b52f992e91b5e079f2881262ab5590090
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195908
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>