MB-58413: Avoid overflow in HHWheelTimerBase<Duration>::scheduleTimeout
Some tasks use INT_MAX as their snooze time, which has been a pattern
in KV for sometime - it really means this task should not periodically
wake, i.e. sleep until explicitly told to run.

Merging of a ASAN enabled folly dependency highlighted that tasks using
the INT_MAX snooze caused an overflow inside the scheduleTimeout
function.

This commit looks for tasks which are in the "sleep until event" state
and skips the call to scheduleTimeout - as there is no timeout for the
task.

ExecutorPoolTest already has code which uses the INT_MAX pattern, e.g.
ExecutorPoolTest::WakeMultiple, and this commit relies on those tests
for validating this commit does not regress. Then remerging the folly
ASAN dependency will give coverage that the issue is resolved.

Change-Id: I1786d4b499becffe01ee477c81bb734343de976d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/196272
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>