MB-58321: Don't register cursor on an empty and closed checkpoint
The unit test identified a regression since 7.2.0 where register cursor
places the cursor too "early", resulting in the cursor iteration seeing
an item it has already seen (and hence the ActiveStream seqno monotonic
violation seen in the MB).

The fix here backports some code from master into registerCursorBySeqno
so that if the startBySeqno matches the end of the current checkpoint,
and more checkpoint's are available, we skip to the next checkpoint.

E.g. given the following scenario.

CP:10 start=1001,end=1001, state=closed (our empty CP)
CP:11 start=1001,end=1001, state=closed (CP actually storing 1001)
  - keys{1001}
CP:12 start=1002,end=1002, state=open
  - keys{1002}

Prior to this fix CP:10 appears to be a candidate by the following code
in registerCursorBySeqno(startBySeqno=1001)

   else if (startBySeqno <= en)

and the cursor is planted at the end of CP:10, meaning it will iterate
into CP:11 and return 1001, a key already seen by the client -
triggering the monotonic failure in DCP.

After this fix, with the addition of the following check.

    const bool isLastCkpt = std::next(ckptIt) == checkpointList.end();
    if (startBySeqno == en && !isLastCkpt) {
        continue;
    }

CP:10 is no longer a candidate, and neither is CP:11. Instead CP:12
is correctly picked and the cursor placed at the start. Now when the
cursor is iterated for items, seqno:1002 is the next item seen.

A further fix is added because the unit test failed because tryBackfill
was incorrectly set to true. This is fixed by conditionally setting
to true only when there is a gap between startBySeqno and the checkpoint
start.

Change-Id: I43673496499393bbce5306394f19b07fa57caf1a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/196388
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>