[test-client] Improve packet validation
Improve packet validation for received packets to make sure
that datatype reflects the actual content in the package

Refactor check for MEMCACHED_UNIT_TEST env variable into
a method to get rid of a warning where a local variable shadowed
a global one. By making it a method which always checks the
environment variable it now works in the context where the
program itself set the environment variable.

Change-Id: I04da0476df17ae01e23843deb9ddf0ee8260b3b5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/196582
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>