MB-58521: Set CAS for lookup operations
As part of tightening the subdoc code a change was made to
only reveal the documents CAS value as part of a successful
operations. The assumption was however not correct as subdoc
allows for multiple "commands" to be executed in a single
operation and for lookups failing one of the paths isn't
a fatal operation.

Given that we've already verified the access to the actual
document we don't leak any information to the client
by adding the CAS in the response in these cases (the
client could request the entire document instead to get
it).

Change-Id: I9536aab3b9da6d2c9ce91ed808c1c9158982a5a0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/196683
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>