MB-58716: CouchKVStoreFileCache: Account memory to 'noBucket'
CouchKVStoreFileCache is a process-wide (cross-bucket) structure,
however it does not correctly switch away from the current bucket when
making calls into it which allocate memory.

As such, mem_used for the bucket which is selected when calls into it
are made can be incorrectly inflated / deflated (depending on if the
call allocates or deallocates memory).

Fix by switching away to "no bucket" when making process-wide memory
(de)allocations.

Change-Id: I8e6bad768a461e32d7aa80f2395c58ec7bf463b7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197325
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>