[Refactor] CheckpointManager: Avoid unnecessary string formatting
In CheckpointManager::addOpenCheckpoint() we unconditionally perform a
bunch of string formatting for the prefix of an exception message _if_
we hit some invalid cases. Formatting this string is not free - it
requires multiple heap allocations (one for the resulting std::string,
plus the sub-strings returned by to_string), plus whatever work fmtlib
needs to perform to format the string.

While CheckpointManager::addOpenCheckpoint() isn't super-hot, we can
create checkpoints at a pretty high rate if the mutation rate is high.

Restructure so we only perform the string formatting if we actually
need to throw an exception.

(Note: Identified when looking for arena mis-accounting, I observed
multiple heap allocations / frees coming from this line).

Change-Id: I1171d0fc035d7a54318a393cbd830b4d14606d40
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197321
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>