MB-58463: Ensure unref checkpoints removed at CM::registerCursorBySeqno
When we re-register a cursor with the same name, the function must
ensure 2 things with regard checkpoint removal:

  1. If the cursor is re-registered into the same checkpoint, then we
     don't want to remove that checkpoint as side effect of the
     re-registering operation. That is what we fixed in in Trinity via
     merge at https://review.couchbase.org/c/kv_engine/+/195643
  2. If the cursor is re-registered into a subsequent checkpoint, and
     if the operation makes the old checkpoint unreferenced and
     eligible for removal, then we must remove that checkpoint in place

In the merge that fixed (1) we introduced a bug that broke (2). We fix
(2) here.
The consequence of breaking (2) is that a vbucket might enter a state
where closed/unreferenced checkpoints are no longer removable from CM,
which might push the node quickly into a unrecoverable OOM state.
See MB for details.

Change-Id: I2724b0ccf94c7f774624d524dbad9061b7ff928d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197028
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>