MB-58609: NIM needs to run in its own thread
Previously we used the "main thead" of the program to accept
clients, reconfigure network interfaces etc. This is problematic
during shutdown as we terminate the event loop for the main
thread and initiate shutdown for the other components. As part
of shutting down we need to stop all of the worker threads,
but we might have a command currently trying to run network
reconfigure. This will hang "forever" as we've stopped running
the event dispatch loop.

Change-Id: Ieebe65e2b542d9878b9faaf01b832c87cc7df094
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197104
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker