MB-58664: Use requested TTL for Set(preserveTTL)
If preserveTTL option is used in conjunction with a specific TTL, then
the resulting TTL should be the existing TTL if the docuemnt exists,
if not then use the request's TTL.

However, when using using Set() to create a document on a
full-eviction bucket with those TTL options (preserveTTL=true,
request.TTL != 0), the requested TTL is incorrectly left as zero.

This is due to how preserve is implemented - it unconditionally
"preserves" the existing StoredValue's TTL, even if the existing
StoredValue is not an alive document. In the case of a non-alive
StoredValue, the TTL is the default of zero and so that is the one
used for the new mutation.

Fix by only copying the TTL if there's actually a valid existing
document.

Change-Id: Ib4a77df9737d4ff9865a6e3874858f64c8afadb1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197179
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>