MB-58733: Reintroduce the CheckpointTest::RegisterDuplicateCursor
Removed at https://review.couchbase.org/c/kv_engine/+/195643
by the assumption that re-registering a cursor cannot trigger
CheckpointRemoval. That assumption is wrong, as in general a cursor
might be re-registered in a checkpoint subsequent to the old position,
thus making the old checkpoint eligible for removal.

The test is updated for covering that case and reintroduced. Original
test added in MB-56094.

Change-Id: Ic17afb73ad5ec86f4d17034a6aa86ac975fb1dba
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197526
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>