MB-58834: Fix memory tracking of EPEngine at start of ctor / end of dtor
+Problem 1: destruction+

When an EpEngine object is destroyed, it fails to account all memory
deallocations, because the ArenaMalloc client is unregistered _inside_
the dtor:

    EventuallyPersistentEngine::~EventuallyPersistentEngine() {
         ...
         // Engine going away, tell ArenaMalloc to unregister
         cb::ArenaMalloc::unregisterClient(arena);
         // Ensure the soon to be invalid engine is no longer in the ObjectRegistry
         ObjectRegistry::onSwitchThread(nullptr);
    }

Because EventuallyPersistentEngine has various sub-objects owned by
unique_ptrs, which are destroyed at the end of the dtor - i.e. after
unregisterClient - these sub-objects are not correctly accounted in
memory tracking.

This results in the memory usage of the arena not being reduced to
zero before unregisterClient is called. While technically benign in a
production system (there's no actual leak of heap, and if the arean is
re-used then the counters are just reset to zero), it makes it hard to
reason about memory usage and ensure that all memory is actually
correctly tracked.

+Problem 2: construction+

There is a similar issue at EpEngine creation - we cannot call
ObjectRegistry::onSwitchThread() until a valid pointer to the new
engine exists, but we do need to account allocations of sub-objects -
which occur within the the ctor.

There is code in place which _attempts_ to handle this situation -
create_ep_engine_instance() calls ArenaMalloc::switchToClient() before
constructing EpEngine, which should be sufficient to ensure that any
heap (de)allocations via operator new/delete are correctly accounted
to the under-creation engine. Unfortunatly there's a flaw -
allocations _will_ only be correctly accounted as long as the current
bucket isn't changed. Specifically, if some other code (executed
during the ctor) calls NoBucketAllocationGuard(), then when saving the
"current" engine in NoBucketAllocationGuard ctor the nullptr will be
stored, which when the NoBucketAllocationGuard dtor runs then the
current engine (and underlying arena) will be switched to the "no
arena" - breaking tracking from that point onward.

+Solution+

What we need to do is call ObjectRegistry::onSwitchThread() as soon as
possible once there's actually an EpEngine* to pass to it -
i.e. before any potential NoBucketAllocationGuard calls. In general we
cannot know whereabout those calls can occur, but _if_ we call
ObjectRegistry::onSwitchThread as part of initialising the first
member variable of EpEngine (and EpEngine has no base class - which is
doesn't) - then we can guarantee that we have installed the correct
EpEngine TLS value before any other code can attempt to read it as
part of NoBucketAllocationGuard.

This is achieved by introducing a new helper class -
EpEngineArenaHelper - which is declared as the first member variable
of EpEngine (so it's ctor runs before any other member variables'
ctor). EpEngineArenaHelper's ctor simply calls
ObjectRegistry::onSwitchThread(engine), swiching the current thread to
the being-created engine. Given the ArenaMallocClient is needed as
part of onSwitchThread(), it is moved from EpEngine to
ArenaMallocClient so initialised before onSwitchThread needs to be
called.

Once this is all in place, EpEngine creation and destruction paths are
updated to not switch away / unregister the client until the EpEngine
dtor has finished. All memory allocated by EpEngine is now correctly
tracked.

Change-Id: I78114d727b881d9311c05a2758dbe36289769a96
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197627
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>