MB-57794: Remove superfluous locking and reduce lock scope
It is no point of locking no-bucket and the current bucket
as no-buckets lifetime exceeds the lifefime of all connections
and we have a ref-count for the current bucket.

Refactor to use BucketManager::forEach which bumps the ref count
for the bucket and perform the callback without holding the lock
for the bucket.

Extend the check to verify the euid provided in the request

Change-Id: I2636a47188fb4d896707ee5a11a93e2bb859adff
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197291
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>