MB-58833: Fix gcc -Wshadow=compatible-local warnings
We have recently hit a nasty issue caused by variable shadowing and the
consequent usage of the wrong variable reference (MB-58463).

The -Wshadow flag would give us a compilation warning and reduce the
chance of making that kind of error.

I start by introducing -Wshadow=compatible-local for GCC in a related
tlm change. That is the loosest of multiple options for that flag.
KV immediately fails the build with a number of warnings that I fix
here.

In subsequent patches I'll make equivalent changes for Clang and MSVC.

Change-Id: I62b30da2b2311e4f0d32d6eca4c3f00a1c85bd0f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197732
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>