MB-55396: Document metrics with fixed labels
For some metrics, we have one definition in stat_definitions.json per
fixed set of labels (for example: kv_ops{op=get, result=hit} gets its
own definition with key=get_hits). This creates a problem for generating
documentation, because while we have all of the possible labels
documented, we don't have a description for the metric family itself
(kv_ops in the example).

This patch allows a description for the metric family itself to be
specified as an object with the keys (family, description). The provided
description is used as the help text for the family in the generated
documentation.

This patch also enforces at compile-time that we have such definitions
for such metrics, so I have added them as part of this change.

Change-Id: I90dc2b12d150c2432da83eac83749edffe36baa3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197756
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>