MB-58834: During EpEngine::destroy() ensure that current engine is not selected
EventuallyPersistentEngine::destroy() destroys the given engine, and
unregisters it with the memory tracking client. As such, it would be
invalid to have thread-local current_engine set to the
about-to-be-destoyed engine. In production code (i.e. daemon calling
EpEngine::destroy at bucket deletion) this expectation is correct -
the 'current' engine is nullptr inside daemon.

However in various unit-test code this is not the case - destroy() is
called when the current engine is still set to it.

To aid in tracking down memory accounting issues (and tests are
behaving in the same way as production), add an expectation to
EpEngine::destroy that the current engine is not the object being
destroyed.

This then highlights a number of problematic unit tests, which need to
be fixed to deselect the engine before destroying it. The vast
majority of unit test uses are via the SynchronousEPEngine subclass;
these can all be addressed by overriding destroy() and switching away
from the current engine before calling the base class destroy(). For
the remaining use-cases (which use create_ep_engine_instance() like
bucketmap), we explicitly switch away from the current engine before
calling destroy().

Change-Id: Ifdbaf99831abd3472c9e02551cacb353f4545ed4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198490
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>