MB-9418: Add Secondary Warmup
What is now known as Primary warmup can be ran to the point that
metadata is loaded but no data - at this point a full-eviction bucket
could be brought online.

This patch adds a second phase of warmup creatively named Secondary.
The secondary phase reuses the Warmup code but jumps over the
metadata phases which the Primary warmup has completed. Secondary
warmup now continues in the "background" to populate the cache.

The intention is that this feature would be used with the following
configuration

    warmup_min_memory_threshold = 0
    warmup_min_items_threshold = 0
    secondary_warmup_min_memory_threshold = 100
    secondary_warmup_min_items_threshold = 100

Change-Id: I35222c2e06c140b45b8369e6b01e62a7f541fe6b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/206444
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>