MB-58649: Improve RangeScanTest stability
MB-58649 adds a new failure path to range-scan-create to better handle
the case where a command finds the vbucket data file is missing (not
yet flushed and is tagged with creation=true).

Switch RangeScanTest to use execute instead of manual send/recv as
execute handles temp failure automatically and removes sporadic failure
where the test was racing with bucket creation.

Change-Id: I1ceff965116b7b015de58c9e1eebba973c39d4d5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197902
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>