Fix sporadic test failure in MaxConnectionPerBucket
Disconnecting a client (and updating the current count of
connections to a bucket) is an async task on the server
and depends on the scheduling.

Make the test more robust by waiting for the count of
connections to drop down to the expected max number of clients
before trying to connect an internal connection.

(Note there is no point of connecting more than a single
connection to verify that internal connections may bypass
the limit)

Change-Id: I7311cec152204957047147de4b78e11c18355043
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/197955
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>