MB-9418: [Refactor] Warmup standardise requests for the shard count
This commit began as "can we remove EPBucket& from Warmup"? For
example Warmup doesn't need to keep asking for the vbMap.shards.size
when it has shard sized containers...

However EPBucket is very much embedded into Warmup and as such the
refactoring just standardises the places which care about shard count.
The refactoring just shortens the levels of indirection to get to such
a value.

Change-Id: Ie4d40f0db3cac825a231308921ac3bd78f909455
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/206093
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>