MB-58834: KVBucketTest: Don't account temp strings to engine
During KVBucketTest::initialize(), a temporary std::string is built up
for the config to use when creating the engine. This string is
allocated outside the engine scope and hence accounted to 'no bucket',
however it goes out of scope and is deallocated when the engine is
selected, creating an allocation mismatch.

Fix by ensuring it goes out of scope when 'no bucket' is selected.

Change-Id: If93667024d3269816c94b528eb03144ba80d2a29
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198492
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>