MB-58894: Set a max limit of 100 000 keys in GetKeys
GetKeys allows the user to specify the number of keys
to return (default to 1000 if no limit is provided by
the client). The problem was howver that no max limit
was specified, so that a client could request up to
4G documents to be returned. The response is built
in memory and consist of a "list" of keys encoded
like:

    |uint16_t key length| length bytes of key |

With no limit one would easily run out of memory.

If more than 100000 keys is needed one would need
to rerun the command with a new "start key"

Change-Id: I5b622cf90c001371d51dcadbb40dba73ff1cc0a1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198015
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>