MB-58896: Catch exception when shutting down audit
mktime() may fail (and return time_t(-1) which would cause the
timestamp to be invalid (and sprintf would overflow causing
an exception to be thrown).

In the call path from a "client" the exception will be caught
in the client context, but in the context of the destructor
no one caught the exception causing the process to crash.

Change-Id: Id8e0aefe57040478351cdab91178cd2ea3998f33
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198034
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>