MB-59022: Set engine correctly for VBucketSyncWriteTimeoutTask
VBucketSyncWriteTimeoutTask - as introduced via MB-39815 in 7.1.0 -
constructs its base class GlobalTask via the ctor taking a Taskable,
not the ctor taking a EventuallyPersistentEngine.

This is problematic as when the task runs, the execution pools do not
switch to the current engine (as GlobalTask::engine is null) - and as
a result any (de)allocations made are not correctly accounted.

Fortunately VBucketSyncWriteTimeoutTask doesn't appear to (de)allocate
any memory when run, so this issue is currently benign, however it
could cause mis-accounting if VBucketSyncWriteTimeoutTask is modified
and starts allocating memory.

Change-Id: I27b359f3e15e379a04e1fcd9be5dc35ebb763eb7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198493
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>