[BP] MB-61464: Verify that new impl to format timestamps
BP of MB-53572 to 7.2.6.

Add unit test to verify that the new implementation
to generate timestamps produce the same output as
the old implementation

Running benchmark on my (8 X 3900 MHz CPU s) Linux box:

    ----------------------------------------------------------------
    Benchmark                           Time         CPU  Iterations
    ----------------------------------------------------------------
    ISOTimeClassBackend/threads:1    3696 ns     3696 ns      189222
    ISOTimeClassBackend/threads:2    3251 ns     4815 ns      144620
    ISOTimeClassBackend/threads:4    3602 ns     5544 ns      126980
    ISOTimeClassBackend/threads:8    3445 ns     6803 ns       80000
    SpdlogBackend/threads:1           883 ns      883 ns      804346
    SpdlogBackend/threads:2           467 ns      935 ns      737104
    SpdlogBackend/threads:4           299 ns     1193 ns      578460
    SpdlogBackend/threads:8           339 ns     2610 ns      267360

Unfortunately there is a problem with the spdlog backend on Mac
that it produce an incorrect timestamp if the provided time point is
some time in the past. This functionality is only used in couchstore
when trying to format the timestamp in the database file headers
(and given that MacOS is only a development platform this might
not be a show-stopper)

Change-Id: I29854e8df39ce1faeba8ec5d352a005a92ea38be
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/209400
Well-Formed: Restriction Checker
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>