MB-58548: Fix markers checks for NonHistory->History jumps
We have some sanity check in ActiveStream::markDiskSnapshot for
ensuring that we always send a disjoint sequence of snapshots. Ie, we
check that the startSeqno of a marker is always greater than the
endSeqno of the previous marker.

That applies always but in the case of NonHistory->History jumps in CDC
setups. In fact, that case generated two subsequent markers that carry
the same snapshot range.

Change-Id: Ica411446fc41c0e886072c2a4faa7c03bfa8cf6d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198303
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>