MB-58961: Remove ActiveStream::lastReadSeqnoUnSnapshotted
That is set to the last seqno that a stream has got from checkpoints
but still not pushed into the stream readyQ.

Used for passing that information down to
ActiveStream::snapshot(<items>) and setting AS::lastReadSeqno. That
doesn't need to be a class member as that information is computed
locally when the stream makes up a snapshot from checkpoint.

Also, lastReadSeqnoUnSnapshotted represents a non-atomic information on
the stream state, as that is updated for every item processed under
streamLock in the middle of the execution of
AS::nextCheckpointItemTask().

Change-Id: Iac36349920203e81ba9898137a902c2d1def2048
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198562
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>