MB-59062 [1/7]: Refactor ItemPager in preparation for EpTask
+Overview+

A number of bugs have been found in the usage of GlobalTask to manage
both per-bucket (ep-engine) tasks and 'global' tasks not tied to a
single bucket.

Specifically, GlobalTask has two different ctors, taking either a
Taskable, or an EpEngine instance. The former (Taskable) ctor is
intended when the Task is not associated with a specific bucket, and
hence per-bucket memory tracking should not be performed. The latter
(EpEngine) ctor is intended to be used for per-bucket tasks, and sets
GlobalTask::engine to a non-null value.

However, it is relatively easy to call the "wrong" ctor - there's been
two instances so far (MB-58576 and MB-59022) where the Tasable ctor
was called for bucket-specific tasks and hence we did failed to track
memory correctly.

Moreover, GlobalTask - while now part of the "server" part of
KV-Engine, still includes ep-engine specific headers, as it needs to
perform memory tracking operations for ep-engine instances.

For these reasons, this sequence of patches restructures GlobalTask,
moving the ep-engine specific functionality into a new EpTask
subclass. GlobalTask::engine is moved to EpTask, along with the
corresponding ctor, making it more difficult to accidently call the
wrong ctor. This also removes all of the "engines/ep" includes from
GlobalTask, creating a cleaner separation of logic.

+Changes in this patch+

This first patch performs some initial refactoring to facilitate the
split of EpTask away from GlobalTask.

We have two sub-classes of ItemPager - StrictQuotaItemPager (the
currently used one for shipping configurations) and
QuotaSharingItemPager (in development for serverless deployments). The
former is a bucket-specific task, while the latter is global,
applicable to multiple buckets:

                     GlobalTask
                          ^
                          |
                   NotifiableTask
                          ^
                          |
                      ItemPager
                          ^
                          |
            .-------------+------------.
            |                          |
    StrictQuotaItemPager     QuotaSharingItemPager

The split of ep-engine specific logic into EpTask means these two
ItemPager tasks can no longer share the same parent class -
StrictQuotaItemPager must inherit from EpTask, QuotaSharingItemPager
from GlobalTask). To keep much of the shared logic in ItemPager,
restructure the hierarchy so ItemPager becomes a top level base class,
and the two concrete ...ItemPager classes inherit from ItemPager, and
NotifiableTask (for now):

                     GlobalTask
                          ^
                          |
           .--------------+--------------.
           |                             |
           |                             |
           |          ItemPager          |
           |              ^              |
      NotifiableTask      |          NotifiableTask
           ^              |              ^
           |    .---------+---------.    |
           |    |                   |    |
    StrictQuotaItemPager     QuotaSharingItemPager

Subsequently when EpTask is introduced, StrictQuotaItemPager will be
re-parented to a new EpNotifiableTask which in turn inherits from
EpTask.

Change-Id: Id5d55261623930b9f0628d2c23f8fad5ec2d2fb0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198771
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>