MB-58961: Remove seqno checks for backfill from AS:registerCursor
Checks originally introduced in MB-22960.
In the current code base those checks are already enforced in
CM::registerCursorBySeqno(), which is were we determine whether a
backfill is needed or not.

That area is in general well covered by unit tests. Specifically for
MB-22960 we have the following original tests that guard this change:

- MB22960_cursor_dropping_data_loss
- MB25056_do_not_set_pendingBackfill_to_true
- MB25056_backfill_not_required

Change-Id: I573059d2d05701e2024b890b1d84d708768e9833
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198617
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>