MB-41352: Prune audit log older than X seconds
If the configuration includes "audit_prune_age":<somenumber> the
audit daemon will remove old audit files as part of checking if
files should be rotated or not.

If the value is set to 0 (or not present in the configuration)
the audit daemon will not try to prune old audit logs.

The file age is determined from the files "last write timestamp",
and all files matching "hostname[*]-audit.log" will be subject
for removal

Change-Id: I34205e514e16beba0248e9b9ac6bfa3ba59a4195
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198638
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>