MB-59062 [7/7]: Move ep-engine logic from GlobalTask to EpTask
Remove all references to EpEngine from GlobalTask, making it only deal
with non-bucket related tasks.

Transfer the relevent logic to GlobalTask.

Remove any ep-engine specific logic from the ExecutorPools (given they
just deal with GlobalTask base class), for handling the accounting of
(de)allocation when tasks are reset, add a new invokeViaTaskable()
method to the Taskable class, which should perform any necessary logic
to run code against a given task - in the case of EpEngineTaskable
this is where we switch to the given engine instance.

Change-Id: Ic2dcb1aca0dd3fbf0453d69fc794f9d9b4954b9b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198805
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>