MB-55027: Abstract away the StatVector from StatsTask
Instead of passing in the StatVector to the sub-class, pass in a
callback. The callback is currently set to simply append to the vector,
which is what the sub-classes used to do manually.

The plan is to keep the StatVector internally, but change the callback
to append_stats to the output every N bytes, so we don't keep large
responses in memory.

Change-Id: I2b344c0ba3ff8cbf9f6ea33fb071a6d0ee59f842
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199276
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>