MB-55027: Send DCP per-stream stats after all other DCP stats
Separate out the per-stream stats from Producer::addStats. Add a new
method called addStreamStats which sends just the per-stream stats to
the front-end.

The per-stream stats can generate a very large output when there are
many streams. This output is currently spooled in the front-end and can
be in the gigabyte range when thousands of streams are present.

The plan is introduce an EWB phase for the dcp stats call. This
naturally requires some sort of state machine to be introduced, but to
avoid introducing states for the entire call, separate out just the
stream stats. The iterations over streams will be made to EWB when some
condition is reached in a future patch.

Note this reorders the output of STAT(dcp). Given this output is not
specified and the contents is still the same, this is probably okay.
Moreover, separating the wall of text that is the per-stream stats from
the producer stats is probably easier to read through.

Change-Id: I730c53c2a26288d3ce30ea1ff5fd9559376571aa
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198809
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>