MB-55027: Allow stats commands to throttle and use for dcp stats
Used throttled to split off dcp stats into two phases, with a yield
in-between sending per-producer stats and per-stream stats.

This implementation will be revised in later patches to yield during
the connmap visit if the connection sendBuffer is full, and the
throttled code handling in the code will be changed to accomodate that.

Change-Id: Ib73180f560a18126662760433907e10f3ab416bb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198811
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>