MB-59079: Fix underflow in ep_pending_compactions
Remove the fragile counting logic which currently causes the counter
to underflow when compaction reschedules (double decrement).

Use the size of the compactionTasks map as a better indicator of how
many compactions are "pending" (this counter has always included
pending/running and this commit makes no change to that).

Note the underflow was reproducible with the addition of the following
EXPECT in CollectionsEraserPersistentWithFailure::FailAndRetry

    EXPECT_EQ(0, engine->getEpStats().pendingCompactions);

The test now just checks the compactionTasks size directly.

Change-Id: I517a1e8f980ae2c6bd5bffc5c6d5471752a3a228
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/198904
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>