MB-58135: Warmup: Ignore corrupted accessLog entries
When loading the access logs during warmup, any read errors
(MutationLog::ReadException) are caught and cause that file to be
skipped; however other errors due to logical corruption of the file
are not. If an invalid magic is found in the access log entry (or
similar logical error), then the thrown std::invalid_argument is not
caught and will terminate the process as it is not caught at any
higher levels.

As such, change to catch any std::exception thrown.

Change-Id: I7cbfdd000a18d1269c886b0636534f0b08561945
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/199027
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker